+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Coinfloor parser: new "Deposit and Withdrawal" data file format added.
 - Ledger Live parser: new data file format added.
 - GateHub parser: new data file format added.
+- Trezor parser: try and get symbol name from filename.
 ### Changed
 - Sort wallet names in audit debug as case-insensitive.
 - Data source names in config are now case-insensitive.